Add mouseClicked events to the control boards
MouseClicked events are added to FloorControlBoards and
ElevatorControlBoard. When the circle representing a button is clicked
the system knows the 'passenger' pressed callbutton or button within
elevator.

Also a new method is added in Passenger class to simulate the pressing
of elevator's floor button.

Also LiftControlBoard is renamed to ElevatorControlBoard.